refactor(test): split files by testing object
